+++ src/backend/executor/executor_merge_pipeline.c
@@ -202,29 +202,29 @@ int handle_merge_with_pipeline(cypher_executor *executor, cypher_query *query,
                 const char *joins_str = sql_builder_get_joins(mctx->unified_builder);
                 const char *where_str = sql_builder_get_where(mctx->unified_builder);
 
-                char id_sql[4096];
-                size_t pos = 0;
-                pos += snprintf(id_sql + pos, sizeof(id_sql) - pos, "SELECT ");
+                sqlite3_str *id_str = sqlite3_str_new(executor->db);
+                sqlite3_str_appendall(id_str, "SELECT ");
                 bool first_col = true;
                 for (int vi = 0; vi < vcount; vi++) {
                     transform_var *tv = transform_var_at(mctx->var_ctx, vi);
                     if (tv && tv->kind == VAR_KIND_NODE) {
-                        if (!first_col) pos += snprintf(id_sql + pos, sizeof(id_sql) - pos, ", ");
-                        pos += snprintf(id_sql + pos, sizeof(id_sql) - pos,
+                        if (!first_col) sqlite3_str_appendf(id_str, ", ");
+                        sqlite3_str_appendf(id_str,
                                         "%s.id AS \"%s_id\"", tv->table_alias, tv->name);
                         first_col = false;
                     }
                 }
                 if (!first_col) {
                     if (from_str && from_str[0])
-                        pos += snprintf(id_sql + pos, sizeof(id_sql) - pos, " FROM %s", from_str);
+                        sqlite3_str_appendf(id_str, " FROM %s", from_str);
                     if (joins_str && joins_str[0])
-                        pos += snprintf(id_sql + pos, sizeof(id_sql) - pos, " %s", joins_str);
+                        sqlite3_str_appendf(id_str, " %s", joins_str);
                     if (where_str && where_str[0])
-                        pos += snprintf(id_sql + pos, sizeof(id_sql) - pos, " WHERE %s", where_str);
+                        sqlite3_str_appendf(id_str, " WHERE %s", where_str);
 
+                    char *id_sql = sqlite3_str_finish(id_str);
                     sqlite3_stmt *match_stmt;
-                    if (sqlite3_prepare_v2(executor->db, id_sql, -1, &match_stmt, NULL) == SQLITE_OK) {
+                    if (id_sql && sqlite3_prepare_v2(executor->db, id_sql, -1, &match_stmt, NULL) == SQLITE_OK) {
                         if (sqlite3_step(match_stmt) == SQLITE_ROW) {
                             int mcols = sqlite3_column_count(match_stmt);
                             for (int mc = 0; mc < mcols; mc++) {
@@ -246,6 +246,9 @@ int handle_merge_with_pipeline(cypher_executor *executor, cypher_query *query,
                         }
                         sqlite3_finalize(match_stmt);
                     }
+                    sqlite3_free(id_sql);
+                } else {
+                    sqlite3_free(sqlite3_str_finish(id_str));
                 }
             }
             cypher_transform_free_context(mctx);